+++ source/extensions/omni.isaac.lab/omni/isaac/lab/sensors/imu/__init__.py
@@ -1,0 +1,16 @@
+# Copyright (c) 2022-2024, The Isaac Lab Project Developers.
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""
+IMU Sensor
+"""
+
+from __future__ import annotations
+
+from .imu import IMU
+from .imu_cfg import IMUCfg
+from .imu_data import IMUData
+
+__all__ = ["IMU", "IMUCfg", "IMUData"]