+++ source/extensions/omni.isaac.lab/docs/CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 ---------
 
+* Added IMU sensor implementation that directly accessess the physx view :class:`omni.isaac.lab.sensors.IMU`. The
+  sensor comes with a configuration class :class:`omni.isaac.lab.sensors.IMUCfg` and data class
+  :class:`omni.isaac.lab.sensors.IMUData`.
+
+
 0.18.6 (2024-07-01)
 ~~~~~~~~~~~~~~~~~~~